requirements.txt
streamlit>=1.35.0
spacy>=3.7.2
pdfplumber>=0.10.3
python-docx>=0.8.11
